feat: add user role management DTOs and responses
- Introduced CreateUserRoleDto and UpdateUserRoleDto for user role creation and updates.
- Added UserRoleDetailResponseDto and UserRoleResponseDto for detailed and summary responses of user roles.
- Implemented pagination response structure with PaginationResponseDto.
- Created project-related response DTOs including ProjectApplicationResponseDto, ProjectDetailResponseDto, ProjectListResponseDto, and ProjectSummaryResponseDto.
- Added JoinProjectResponseDto for handling project join applications.
- Developed ReviewApplicationDto for reviewing project applications.
- Updated ProjectsCommandController to handle new DTOs and responses.
- Refactored existing project query handlers to utilize new response structures.
- Enhanced project utility functions to transform project data into the new response format.
