+++ src/projects/commands/update-project/update-project.handler.ts
@@ -7,9 +7,9 @@ import {
 } from '@nestjs/common';
 import { PrismaService } from '../../../common/prisma/prisma.service';
 import { UpdateProjectCommand } from './update-project.command';
-import { ProjectEntity } from '../../types/project.types';
 import { ProjectUtils } from '../../utils';
 import { Prisma } from '../../../../generated/prisma';
+import { ProjectWithOwner } from 'src/projects/types/project.types';
 
 @Injectable()
 @CommandHandler(UpdateProjectCommand)
@@ -21,7 +21,7 @@ export class UpdateProjectHandler
     private readonly projectUtils: ProjectUtils,
   ) {}
 
-  async execute(command: UpdateProjectCommand): Promise<ProjectEntity> {
+  async execute(command: UpdateProjectCommand): Promise<ProjectWithOwner> {
     const { id, updateProjectDto, userId } = command;
 
     // Check if project exists and user is owner
@@ -54,31 +54,108 @@ export class UpdateProjectHandler
           },
         });
 
-        // If roles are provided, update them
+        // If roles are provided, perform differential update
         if (roles && roles.length > 0) {
           // Validate role IDs using utility function
           const roleIds = roles.map((role) => role.roleId);
           await this.projectUtils.validateRoleIds(roleIds);
 
-          // Remove existing roles and create new ones
-          await tx.projectRole.deleteMany({
+          // Get existing project roles
+          const existingRoles = await tx.projectRole.findMany({
             where: { projectId: id },
+            select: {
+              id: true,
+              roleId: true,
+              maxMembers: true,
+              requirements: true,
+            },
           });
 
-          await tx.projectRole.createMany({
-            data: roles.map((role) => ({
-              projectId: id,
-              roleId: role.roleId,
-              maxMembers: role.maxMembers,
-              requirements: role.requirements,
-            })),
-          });
+          // Create maps for efficient comparison
+          const existingRoleMap = new Map(
+            existingRoles.map((role) => [role.roleId, role]),
+          );
+          const newRoleMap = new Map(roles.map((role) => [role.roleId, role]));
+
+          // Identify roles to update, create, and delete
+          const rolesToUpdate = [];
+          const rolesToCreate = [];
+          const roleIdsToDelete = [];
+
+          // Check for updates and new roles
+          for (const newRole of roles) {
+            const existingRole = existingRoleMap.get(newRole.roleId);
+
+            if (existingRole) {
+              // Check if role needs updating
+              const hasChanges =
+                existingRole.maxMembers !== (newRole.maxMembers || 1) ||
+                existingRole.requirements !== (newRole.requirements || null);
+
+              if (hasChanges) {
+                rolesToUpdate.push({
+                  id: existingRole.id,
+                  data: {
+                    maxMembers: newRole.maxMembers,
+                    requirements: newRole.requirements,
+                  },
+                });
+              }
+            } else {
+              // New role to create
+              rolesToCreate.push({
+                projectId: id,
+                roleId: newRole.roleId,
+                maxMembers: newRole.maxMembers,
+                requirements: newRole.requirements,
+              });
+            }
+          }
+
+          // Check for roles to delete
+          for (const existingRole of existingRoles) {
+            if (!newRoleMap.has(existingRole.roleId)) {
+              roleIdsToDelete.push(existingRole.id);
+            }
+          }
+
+          // Execute differential operations
+          await Promise.all([
+            // Update existing roles
+            ...rolesToUpdate.map((roleUpdate) =>
+              tx.projectRole.update({
+                where: { id: roleUpdate.id },
+                data: roleUpdate.data,
+              }),
+            ),
+            // Create new roles
+            ...(rolesToCreate.length > 0
+              ? [tx.projectRole.createMany({ data: rolesToCreate })]
+              : []),
+            // Delete removed roles
+            ...(roleIdsToDelete.length > 0
+              ? [
+                  tx.projectRole.deleteMany({
+                    where: { id: { in: roleIdsToDelete } },
+                  }),
+                ]
+              : []),
+          ]);
         }
 
         return project;
       });
 
-      return this.projectUtils.getProjectWithDetails(updatedProject.id);
+      const projectWithDetails = await this.projectUtils.getProjectWithDetails(
+        updatedProject.id,
+      );
+
+      if (!projectWithDetails) {
+        throw new BadRequestException(
+          'Failed to retrieve updated project details',
+        );
+      }
+      return projectWithDetails;
     } catch (error) {
       if (error instanceof Prisma.PrismaClientKnownRequestError) {
         throw new BadRequestException(
