+++ src/projects/dto/application-decision.enum.ts
@@ -1,0 +1,4 @@
+export enum ApplicationDecision {
+  APPROVED = 'APPROVED',
+  REJECTED = 'REJECTED',
+}
