+++ src/projects/dto/index.ts
@@ -1,8 +1,28 @@
 // Request DTOs
 export * from './create-project.dto';
+export * from './create-project-role.dto';
 export * from './join-project.dto';
 export * from './update-project.dto';
+export * from './update-project-status.dto';
+export * from './review-application.dto';
 export * from './user-role.dto';
+export * from './create-user-role.dto';
+export * from './update-user-role.dto';
 
 // Response DTOs
 export * from './project-response.dto';
+export * from './pagination-response.dto';
+export * from './user-role-response.dto';
+export * from './user-role-detail-response.dto';
+export * from './project-owner-response.dto';
+export * from './project-role-response.dto';
+export * from './project-summary-response.dto';
+export * from './project-member-response.dto';
+export * from './project-application-response.dto';
+export * from './project-detail-response.dto';
+export * from './project-list-response.dto';
+export * from './create-project-response.dto';
+export * from './join-project-response.dto';
+
+// Enums
+export * from './application-decision.enum';
