+++ src/projects/dto/pagination-response.dto.ts
@@ -1,0 +1,11 @@
+/**
+ * Pagination information for list responses.
+ */
+export interface PaginationResponseDto {
+  page: number;
+  limit: number;
+  total: number;
+  totalPages: number;
+  hasNext: boolean;
+  hasPrev: boolean;
+}
