+++ src/projects/dto/update-project-status.dto.ts
@@ -1,0 +1,13 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsEnum } from 'class-validator';
+import { ProjectStatus } from '../../../generated/prisma';
+
+export class UpdateProjectStatusDto {
+  @ApiProperty({
+    description: 'New project status',
+    example: ProjectStatus.IN_PROGRESS,
+    enum: ProjectStatus,
+  })
+  @IsEnum(ProjectStatus)
+  status: ProjectStatus;
+}
