+++ src/projects/queries/find-all-projects/find-all-projects.query.ts
@@ -1,5 +1,11 @@
 import { ProjectFilters, PaginationParams } from '../../types/project.types';
 
+/**
+ * Query object for retrieving all projects with optional filtering and pagination.
+ *
+ * @param filters - Criteria to filter the list of projects.
+ * @param pagination - Parameters to control pagination of the results.
+ */
 export class FindAllProjectsQuery {
   constructor(
     public readonly filters: ProjectFilters = {},
